Allow sharing rocket-specific photos
